Move base schema to runtime::datafusion::schema
